Overhaul frame loading for independent 300x100 and 300x250.
This changeset implements a new cross-frame communication protocol with
Adzerk which enables us to deliver the 300x100 and 300x250 units
independently. Adzerk passes in a comma-separated list of ad frames to
show. Any frames not included in the list are hidden to regain their
space.

In this reshuffle, care has been taken to reuse existing iframe ids so
that the Adblock Plus whitelist does not require updating.